api/requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
pillow==10.4.0
numpy==2.3.0
rembg==2.0.0
python-multipart==0.0.9
supabase==2.7.4
python-dotenv==1.0.1